pratica03/automatos-python/requirements.txt
automata_python==0.0.5
